Scoped slugs
This commit introduces some noticeable changes.

First of all it's scoped slugs. This feature make it possible to limit slug
uniqueness to some scope. Scoping works pretty much the same as it does for
validations.

Second is that now uniqueness is not enforced on db level. Now it's done in
application with dm-validations. Therefore, new dependency.